requirements.txt
python-telegram-bot==21.4
youtube-dl==2021.12.17
yt-dlp==2024.8.1